Week_2_airflow/requirements.txt
apache-airflow-providers-google
pyarrow
pandas
sqlalchemy
psycopg2-binary
requests